// packages/user-connectrpc-server/src/generated/chat3_user_connect.ts
// @generated by protoc-gen-connect-es v1.7.0 with parameter "target=ts"
// @generated from file chat3_user.proto (package chat3.user, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import { GetDialogMessagesRequest, GetDialogMessagesResponse, GetUserDialogsRequest, GetUserDialogsResponse, SendMessageRequest, SendMessageResponse, SubscribeUpdatesRequest, Update } from "./chat3_user_pb.js";
import { MethodKind } from "@bufbuild/protobuf";

/**
 * Сервис для работы с пользователями
 *
 * @generated from service chat3.user.Chat3UserService
 */
export const Chat3UserService = {
  typeName: "chat3.user.Chat3UserService",
  methods: {
    /**
     * Получить диалоги пользователя
     *
     * @generated from rpc chat3.user.Chat3UserService.GetUserDialogs
     */
    getUserDialogs: {
      name: "GetUserDialogs",
      I: GetUserDialogsRequest,
      O: GetUserDialogsResponse,
      kind: MethodKind.Unary,
    },
    /**
     * Получить сообщения диалога
     *
     * @generated from rpc chat3.user.Chat3UserService.GetDialogMessages
     */
    getDialogMessages: {
      name: "GetDialogMessages",
      I: GetDialogMessagesRequest,
      O: GetDialogMessagesResponse,
      kind: MethodKind.Unary,
    },
    /**
     * Отправить сообщение
     *
     * @generated from rpc chat3.user.Chat3UserService.SendMessage
     */
    sendMessage: {
      name: "SendMessage",
      I: SendMessageRequest,
      O: SendMessageResponse,
      kind: MethodKind.Unary,
    },
    /**
     * Подписка на обновления (server streaming)
     * Первое сообщение - Update с eventType="connection.established" и connId в data
     *
     * @generated from rpc chat3.user.Chat3UserService.SubscribeUpdates
     */
    subscribeUpdates: {
      name: "SubscribeUpdates",
      I: SubscribeUpdatesRequest,
      O: Update,
      kind: MethodKind.ServerStreaming,
    },
  }
} as const;